Bugfix in converting list to array

+++ get_number_of_collisions.py
@@ -157,13 +157,13 @@ def extract_data_smash(ifile):
 
             had_prop = np.zeros(7,dtype=np.int32)
             pid = stuff_N[9] # we have already read the first line after the interaction event header to get the collision time
-            tot4m = np.array(float(stuff_N[4:8]))
+            tot4m = np.array(np.float64(stuff_N[4:8]))
             had_prop += get_hadron_info_smash(pid) #the function returns a list which is automatically converted into np array
             if n_incoming > 1:
                 for had in range(n_incoming-1):
                    stuff_N = ifile.readline().split()
                    pid = stuff_N[9]
-                   tot4m += np.array(float(stuff_N[4:8]))
+                   tot4m += np.array(np.float64(stuff_N[4:8]))
                    had_prop += get_hadron_info_smash(pid) #the function returns a list which is automatically converted into np array
 
             count_based_on_hadron_property(had_prop,h)
